Make mouse support conditional, quiet zsh source warnings
Trap stderr on `which` for some applications I don't have installed on
everything, because otherwise it shows up in terminal output when I log
in. Also, make mouse support conditional on that support actually
existing, because Cent 7 claims it's compiled with +mouse but still
manages to fail at it somehow.